requirements.txt
Django==2.2.13
mail-parser==3.7.1
python-dateutil==2.7.5
